send_to_ryan_prem/fullstack-app/fast-api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
psycopg2-binary>=2.9.1
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2